+++ simvex/src/main/java/dochigosum/simvex/domain/drawing/presentation/dto/request/PartTransformRequest.java
@@ -3,10 +3,10 @@ package dochigosum.simvex.domain.drawing.presentation.dto.request;
 public record PartTransformRequest(
         Long partId,
         Long drawingId,
-        Integer xCoordinate,
-        Integer yCoordinate,
-        Integer zCoordinate,
-        Integer xRotation,
-        Integer yRotation,
-        Integer zRotation
+        Double xCoordinate,
+        Double yCoordinate,
+        Double zCoordinate,
+        Double xRotation,
+        Double yRotation,
+        Double zRotation
 ) {}
