+++ simvex/src/main/java/dochigosum/simvex/domain/project/presentation/dto/request/PartStoreRequest.java
@@ -2,11 +2,11 @@ package dochigosum.simvex.domain.project.presentation.dto.request;
 
 public record PartStoreRequest(
         Long id,
-        Integer xCoordinate,
-        Integer yCoordinate,
-        Integer zCoordinate,
-        Integer xRotation,
-        Integer yRotation,
-        Integer zRotation
+        Double xCoordinate,
+        Double yCoordinate,
+        Double zCoordinate,
+        Double xRotation,
+        Double yRotation,
+        Double zRotation
 ) {
 }
